requirements.audio.txt
numpy
soundfile
requests
openai-whisper